requirements.txt
edgerun-skippy-core>=0.2.0.dev1
edgerun-faas>=0.0.1.dev3
kubernetes>=18.20.0
